Added an additional unit tests file for matching.py and moved one of the tests from unit_tests_features.py into it. Also added a test for proper equality checking to ensure that IntervalSelectors are properly hashable.
Also updated the CI config for recent new testfiles and removed a template/unnecessary function from unit_tests_statistics.py
